remove mandating mechanism
this implemenation actually did not work. We don't assume funders to mandate retroactively, but only on their own grants.

the "correct" way to mandate now is to set incentives to 1, which gives stronger rewards to those sharing.
However, this is not a hard mandate, in a sense that those who receive funding need to share data