simplify and speed up code
The initial issue here was the final [ - 0.5 ] on line 195. This would have been a bug, because it should be 0.5. However, this value never made its way into calculations, because the share-data procedure does not consider descriptive norms when there is no network, and in all network cases all agents at least have one neighbour.

Overall: this does not change any of the results, and should speed up the calculations.